+++ src/components/index.ts
@@ -1,0 +1,6 @@
+export * from './Controls';
+export * from './MediaControls';
+export * from './Slider';
+export * from './Toolbar';
+export * from './MediaControls.style';
+export {default as MediaControls} from './MediaControls'
