+++ src/index.tsx
@@ -1,4 +1,4 @@
-import MediaControls from "./MediaControls";
+import { MediaControls } from "./components";
 import { PLAYER_STATES } from "./constants/playerStates";
 
 export { PLAYER_STATES };
